requirements.txt
absl-py==0.11.0
altgraph==0.17
anaconda==0.0.1.1
aniso8601==8.0.0
app==0.0.1
appdirs==1.4.4
asgiref==3.3.1
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
attrs==21.2.0
auto-py-to-exe==2.7.8
bcrypt==3.1.7
bottle==0.12.18
bottle-websocket==0.2.9
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
config==0.5.0.post0
connexion==2.7.0
crypto==1.4.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
docutils==0.16
EasyProcess==0.3
Eel==0.12.4
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
future==0.18.2
gast==0.2.2
geographiclib==1.50
geolocation-python==0.2.2
geopy==2.0.0
gevent==20.6.2
gevent-websocket==0.10.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.35.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==3.4.0
inflection==0.5.1
isodate==0.6.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.0
matplotlib==3.3.3
maxminddb==1.5.4
maxminddb-geolite2==2018.703
mccabe==0.6.1
mock==4.0.3
Naked==0.1.31
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
opencv-contrib-python==4.4.0.46
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.4
paramiko==2.7.1
pefile==2019.4.18
Pillow==8.1.0
prompt==0.4.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
pycryptodomex==3.9.7
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyVirtualDisplay==1.3.2
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.6.0
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.4.1
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2020.11.18
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.1
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
zope.event==4.4
zope.interface==5.1.0
